Added a method for equipping items to a character for only a specific client.
This is to avoid a race condition where world items would spawn and then the auth server tries to equip them to a character. If the world item had not spawned before the equip item message came in, the item wouldn't properly equip to a player.

This method facilitates equipping of an item to only a specific client.

Client storage is used for linking a player id to the connection id.